% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert dataframe to mcmc.list function.R
\name{df_to_mcmc_list}
\alias{df_to_mcmc_list}
\title{Converts a dataframe to an mcmc.list object}
\usage{
df_to_mcmc_list(df, chain_col = "Chain.number", iter_col = "Iteration.number")
}
\arguments{
\item{df}{The dataframe to be converted to an mcmc.list object}

\item{chain_col}{Do not alter}

\item{iter_col}{Do not alter}
}
\value{
An mcmc.list object
}
\description{
This is an internal function, and is not intended to be called by users directly.
}
