% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcplot modified function.R
\name{mcmcplot.mod}
\alias{mcmcplot.mod}
\title{Plotting function mcmcplot() from mcmcplots() package modified to not automatically open the html}
\usage{
mcmcplot.mod(...)
}
\arguments{
\item{See}{mcmcplot() from mcmcplots() package.}
}
\value{
See mcmcplot() from mcmcplots() package.
}
\description{
Plotting function mcmcplot() from mcmcplots() package modified to not automatically open the html
}
\examples{
See mcmcplot() from mcmcplots() package.
}
