% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcplot modified function.R
\name{mcmcplot.mod}
\alias{mcmcplot.mod}
\title{Plot MCMC draws}
\usage{
mcmcplot.mod(...)
}
\arguments{
\item{See}{mcmcplot() from mcmcplots() package.}
}
\value{
See mcmcplot() from mcmcplots() package.
}
\description{
A modified version of the mcmcplot() function from the mcmcplots. Modified to not automatically open the html file that is created. This version creates and saves various plots and an HTML file collecting them (but does not automatically open the HTML file.) This function is not intended to be called by users directly.
}
\seealso{
See mcmcplot() from mcmcplots() package.
}
